Add rake task to redirect world location news pages
Similar to the last commit, we need to update the publishing platform’s state with how we’re already routing these pages within Whitehall (otherwise entries appear twice in Search).

We need to get a list of affected content items from the content store, as it’s not possible to retrieve the content_ids from within Whitehall.