Add SRI attributes to JS and CSS assets
Many of the other assets used by whitehall come from static, so this only
gives us SRI on the assets actually delivered by whitehall.  The rest will
come when we add SRI to static.  We don't add these attributes to any of
the IE stylesheets because no version of IE currently supports SRI so
there's no point spending the time to calculate the values.