Moves needs assignment to the document level
Following the switch to getting needs from the Publishing API instead of the Needs API, we are now deprecating the use of the needs_id column in the Editions table. We're also associating needs with the content_id of the document, so it makes sense to move that functionality at the document level wherever possible to keep things cleanly separated. Previously this was being done at the Edition level.

There is no show/edit page for documents in Whitehall, so the "Edit needs" form can be reached by clicking the "Add/remove needs" button on the Edition show page.
This form had previously been on the Edition edit page, but has now been moved in order to avoid having to create a new draft every time you want to add/remove needs.

This was decided because we want to bypass the whole edition workflow of creating a draft, publishing etc. The Edition show page was the only viable page to place the form for this purpose.