Use correct republish method
We can't use DataHygiene::PublishingApiRepublisher on news articles
as they are editions. I thought this meant you would have to pass
the documents, but this is not the case. Actually this method is
designed for models which have not yet transitions to being
editions.

Instead, the correct method to use is
'Whitehall::PublishingApi.republish_document_async', an example of
which is used in the 'Dependable' class.

Ideally, world location news articles should be dependencies of
world locations, but this feels like a much larger change, whereas
this PR is just to fix an issue where the default image of world
locations is not updated in news articles if it changes.