Remove legacy Speech type ‘ImportedAwaitingType’
This SpeechType was added in 2012 (https://github.com/alphagov/whitehall/commit/bdbc7ac299cc5b73924c235b8f13bcb814b61345) for the purpose of importing Speeches from an outside source and is fine to remove now. We do not ‘import’ Speeches anymore and there are no Speeches left in the prod db with this SpeechType, so all of this is redundant code.

Checked the db with:

```
Speech.all.map { |s| s.speech_type.singular_name }.uniq.sort
=> ["Authored article", "Draft text", "Oral statement to Parliament", "Speaking notes", "Transcript", "Written statement to Parliament"]
```