Make sure attachments always use the public asset host
Attachments are always served by the whitehall-backend machines because
the NFS mount they are stored on is not available to the
whitehall-frontend machines.  This means any rendered page (such as CSV
previews) is rendered with by that rails app, and gets the admin assets
host.  The CORS settings for the admin asset hosts means that the browser
will reject those assets as not being available to pages on the public
site and so styling and functionality breaks.  We change the asset_host
for production.rb to use a proc that makes sure to always use the public
asset host for any attachments paths (those with system/uploads in the
path) and this should fix the problem.