Switch to a more reliable way of getting draft taxons
Rather than relying on equality of the items within the arrays,
explicitly reject any items which appear in the list of published
taxons, matching the taxons up using the content id.

This fixes an issue where the Childcare and Parenting taxonomy branch
appears as both draft and live within the Whitehall interface.