Add migration to send missing people to rummager
Some people listed in Whitehall have not been indexed in rummager. This
manifests in some finder screens as rows in the people drop down that
have no description which displays as blank or empty. It also means that
searching for one of the missing people at `/government/people` returns
no result even though that person exists.

This migration updates these missing people in the search index.