Rescue from PublishingAPI get links 404 response
There is a race condition in the WhitehallAdmin UI when creating
new editions. When a new edition is saved for the first time, it's
sent to the PublishingAPI asynchronously. At the same time, the
save operation is considered successful by the
Admin:EditionsController and redirects to the show page for that
edition. The show page aims to displays needs information which
it gets from a `get_links` query to the PublishingAPI. It's possible
(I ran into this in my dev environment) that this query is made
before the Document exists in the PublishingAPI and would
therefore return a 404. This results in an error screen being shown
in the AdminUI.

We add a rescue statement to guard against this, prefering to
return an empty needs response as a way of failing gracefully.
A refresh of the show page should render all the correct data at a
later point...