Fix PDF thumbnail timeout
Using `Open3` inside a `Timeout` doesn’t quite work as expected, and
will not actually time out.

This is because `Open3` has an `ensure` block that catches the raised
`Timeout::Error`, and renders the `Timeout` useless.

For more information, see: https://bugs.ruby-lang.org/issues/4681

This change removes the use of `Open3`, which means that the `Timeout`
will actually time out in time for a user to not receive a `504` error.