Add people filtering to MySQL
Non-english locales query the local mysql database, not rummager.
We need to add filtering to people in mysql to make sure that the user
experience is consistent across regions.

Note that whitehall has two different ways to attach people to editions.
The first is simply using the `editions.role_appointment_id` field on
the `editions` table. This is used for `Speech` and allows single role
appointment to be attached to the edition. The second is through the
`edition_role_appointments` table which allows a many to many between
`editions` and `role_appointments`. This is used for `NewsArticle`,
`Consultation`, `Publication` and `FatalityNotice`.