Return empty array when need ids are nil
When editing a document, we display information about any associated
user needs. The code in the view expects a document's `associated_needs`
to respond to `any?`. If a document's `associated_needs` method returns
a `nil`, it errors and the document cannot be edited from
whitehall-admin.

This commit ensures an empty array is returned in this instance.

Why this happens:
Needs ids are exposed through a call to publishing_api's `get_links` and
also through `get_expanded_links`. Our method to return the
`associated_needs` first checks that `get_links` returns the user needs
ids and if so, it then call `get_expanded_links` and uses the response
to display the user needs in whitehall-admin.

When a document is associated with a need that is unpublished,
`get_links` will still return the need id, but `get_expanded_links` will
return a nil. When this happens, our `associated_needs` currently returns
`nil` instead of a duck that responds to `any?`.