Add CarrierWave AssetManagerStorage engine
The `store!` method returns a file as that's what both the `File` and
`Fog` storage engines do in CarrierWave.

I'm raising an exception in the `#retrieve!` method so that we fail fast
if it's ever called. We're only currently interested in saving files to
Asset Manager and not reading them from it.

We have to pass an instance of `File` (instead of an instance of
`CarrierWave::SanitizedFile`) to `#create_whitehall_asset` because
`RestClient` (that we're using in the gds-api-adapters) expects to be
able to call `File#read` with the optional length argument[1].

[1]: http://ruby-doc.org/core-2.4.2/IO.html#method-i-read