Remove specific whitehall admin asset host in dev config
Forcing the asset host to `Whitehall.admin_root` in development config
causes CORS errors when browsing pages rendered by whitehall-frontend.
The affects loading of assets - so requests for images and javascript
resources are not honoured.

Removing the specific setting of the asset_host to
"http://whitehall-admin.dev.gov.uk" means that assets will, by default,
be referenced with relative paths which will allow them to load.