Add a `use_asset_manager` feature flag
When this feature flag is enabled (via the USE_ASSET_MANAGER
environment variable) the
AssetManagerAndQuarantinedFileStorage (used for non-attachment
assets) will upload assets to Asset Manager (as well as store them on
the file system as usual).

Defaulting this feature flag to OFF allows us to merge and deploy
these changes and add the remaining features required as
branches off master.