Publish statistics announcements with a ‘minor’ update_type
We don’t provide workflow on these editing pages.
As soon as the editor saves, the page is
published. If we send ‘major’ update_types it
means email subscribers get spammed and those
pages appear in digests multiple times.

We could prevent emails for
statistics_announcements in email-alert-service,
but that would mean hardcoding a rule for that
document_type. I think this is a better solution
as it keeps this knowledge inside Whitehall.