Copy file to a temporary location before uploading to asset manager
When the AssetManagerStorage#store! method returns to the uploader,
the file in CarrierWave's `cache_dir` is deleted. If we want to
perform the upload asynchronously in the Sidekiq worker, the file could
already be gone by the time the worker runs.

We copy the file to a temporary location (the directory itself will be
created on all environments in a separate change to `govuk-puppet`)
and pass this path to the worker. Functionality will be added to
`AssetManagerWorker` to remove this file later.