Add unique id to path for temporary asset manager files
Before this commit there was a possible race condition where two
assets uploaded at the same time with the same name would be written
to the same `Whitehall.asset_manager_tmp_dir` directory. By adding a
UUID to the temporary path we avoid this possibility.

CarrierWave by default does something similar using the ID of the
CarrierWave model. In this case we want to isolate the use of this
temporary directory to the AssetManagerStorage class and we don't have
access to a suitable model id.