Save the current_user as an author when withdrawing
Currently a user that withdraws a document is not considered to have
been involved in the authoring of that document. Only users that
have made an edit to the draft of the document are added to the `authors`
collection.

The result of this is that a user that makes a small typo correction is
considered an author but a user that withdraws it, which is a much
more significant change to the document is not.

When an editor searches for documents that they have authored that
are in a withdrawn state the search results are then inconsistent as
some documents that they withdrew (and also 'authored' by the current
definition) will be returned but those that they only withdrew but
didn't 'author' are not.

The `EditionWithdrawer` class now supports a `user` argument that adds
the user to the `authors` collection. This commit updates the
`EditionUnpublishingController` to send the `current_user` when a
document is withdrawn.