Remove SRI from admin JS
We've received a number of support tickets relating to issues with the
pop-up menus in Whitehall admin not popping up. These seem to be caused
by the admin JS not loading for some users since we added the SRI
attributes.

We need to make the JS dependent functionality more resilient but in the
short term this commit removes the attributes from the JS.

A ticket will be added to the Whitehall support backlog to review the
use of JS in the Whitehall admin interface and make changes where
necessary before reverting this commit.