Add integration test for asset manager
This is a happy-path test for uploading assets to asset manager. We're
currently using the Organisation model but this could be replaced with any
other model that mounts an uploader.

This should give us some confidence that we haven't broken anything as we make
changes to `AssetManagerAndQuarantinedFileStorage` and also gives us a place to
add a similar test for removing assets.