Scope worldwide organisation roles by ‘occupied’ only
https://trello.com/c/n19MlHrj/239-user-trying-to-add-a-role-to-embassy-page

There’s currently a bug where we’re not showing a
person assigned to a role because a role
appointment that has ended is being picked out
instead which is then being hidden by the view.

This is the problematic page:
https://www.gov.uk/world/organisations/british-embassy-managua

We also wrote a script to check the correctness of
the results returned from the primary/secondary
methods. Its output is now showing no errors:
https://gist.github.com/tuzz/a4923f4248547d7f6b809f2b9dc6fce4