Use assert_nil in searchable_test
To avoid:

DEPRECATED: Use assert_nil if expecting nil

In both of these tests, `#most_recent_change_note` returns `nil`. This
is because `minor_change?` is `false` and `first_published_version?` is
`true`. I'm going to assume that the test set-up is correct and that
`#most_recent_change_note` should indeed by `nil`, and have updated the
assertion accordingly.

In the second test, `#public_timestamp` returns `nil` because
`first_published_version?` is `true` and `first_public_at` is `nil`.
Again I'm going to assume that the set-up is correct and that this
is therefore safe to convert to an assert_nil.