Introduce MigrateAssetsToAssetManager
This class synchronously calls
`Services.asset_manager.create_whitehall_asset` for every Organisation
logo stored under the `White.clean_uploads_root` directory.

We decided to use the files on disk rather than references to
logos on the Organisation database model to determine which files to
send to asset manager. At the moment every file on disk is available
at a public-facing URL irrespective of whether there are still
references to it in the database. It is safest therefore to create
assets in Asset Manager for everything.

Before this can be used we will need to set the legacy url path, and
two cache control headers. We may also want to make the call to
`create_whitehall_asset` asynchronously. These will be handled subsequently.