Delete previously gone coming soon content items
These content items were previously unpublished in https://github.com/alphagov/whitehall/pull/3483 with a 410 Gone status. This commit deletes them completely from the content-store so that they return a 404 Not Found error instead. This is because these items referred to content that was never published, therefore it doesn’t make sense to show a 410 error.