Call create_whitehall_asset unless asset already in Asset Manager
We don't need to create an asset in asset manager if one already
exists at the given legacy_url_path. This will be the case for any
organisation logos that have already been added to whitehall before
this migration is performed, or if the migration code is run multiple
times.