Use most recent feature for speech image
Speeches use the image of the speaker in the sidebar unless the speech
has been featured. If it has then the speech will use the image from the
feature. This isn't particularly obvious but users get further confused
when they feature a speech for a second time in an attempt to change the
image and it doesn't update.

This is because once it has been featured we were always using the image
from the first feature. This commit updates that functionality to use
the most recent feature which gives users the ability to update the
image albeit by a somewhat circuitous route.