Use Asset Manager to get URL of asset
This will ensure that once we switch an uploader from using
`AssetManagerAndQuarantinedFileStorage` to `AssetManagerStorage` we'll
start rendering the assets in whitehall-admin from Asset Manager instead
of the file system.

I'm not entirely sure about making this request to the Asset Manager API
to get the URL. It feels correct to do this because it means that
Whitehall doesn't need to know anything about how the URLs are created.
On the other hand, I wonder whether the additional network request is
overkill given that we know the path (which is `/government/uploads/ +
@legacy_url_path`) and all we need to prepend is the assets host (e.g.
assets-origin.)