Avoid sending empty files to Asset Manager
The call to `File#read` in `OrganisationLogoFile#legacy_etag` meant that
the subsequent call to `#read`, as part of sending the file to Asset
Manager, returned an empty string. This resulted in empty files being
stored in Asset Manager.

I've changed the test so that it checks the content of the file rather
than just the path. This test failed before adding the fix in this
commit so I'm confident that it'll prevent us from breaking this again
in future.