Generate legacy etags according to the algorithm used by Nginx
So that requests for these assets from Asset Manager respond with the
same ETag as requests for the same assets from Whitehall. This ensures
that we don't invalidate caches en masse when switching the serving of
these assets from Whitehall to Asset Manager.

I've copied the tests and implementation of this behaviour from Asset
Manager[1].

I checked the generation of the `legacy_etag` by comparing it to the
value returned by Nginx when requesting the same file:

```
$ curl
http://whitehall-admin.dev.gov.uk/government/uploads/system/uploads/organisation/logo/1/uk-space-agency.jpg
-v > /dev/null
< ETag: "59f0a734-13767"

$ rails c
> path =
> '/var/govuk/whitehall/clean-uploads/system/uploads/organisation/logo/1/uk-space-agency.jpg'
> file = MigrateAssetsToAssetManager::OrganisationLogoFile.new(path)
> file.legacy_etag
=> "59f0a734-13767"
```

[1]:
https://github.com/alphagov/asset-manager/commit/ec119eb91ae3d82130512a6e824a62a8d56fb44b