Data migration to fix broken statistic announcements
Off the back of: https://github.com/alphagov/whitehall/pull/3516

There are a large number of statistics announcements that are
incorrectly returning the announcement page when they should be
redirecting to the statistics. This task attempts to correct all of
these. Due to the number of them though it's difficult to verify all of
them.