Use AssetManagerStorage for Organisation logos
Organisation logos are now being stored in and served from Asset
Manager, so there is no longer any need to store them on the local,
NFS filesystem. This commit switches the storage engine for
`LogoUploader` to `AssetManagerStorage`.

To ensure that the correct file url is presented in the
`OrganisationPresenterTest` "presents an organisation’s custom logo"
test we return an instance of `AssetManagerStorage::File` from
`AssetManagerStorage#store!` in the same way that we do from
`AssetManagerStorage#retrieve!`.

We have added `AssetManagerStorage::File#filename` because a test in
`Admin::OrganisationsControllerTest` calls it directly in one of its
assertions.

In switching to use `AssetManagerStorage` all virus scanning will now
happen in Asset Manager so we no longer need to simulate the virus
scan of organisations logos in the tests.