ConsultationResponseFormUploader uses asset_manager_and_quarantined_file_storage
This commit switches the ConsultationResponseFormUploader to use the
`asset_manager_and_quarantined_file_storage` engine for storage -
meaning that assets will be stored on disk as before, but also sent to
Asset Manager.

We made a similar change to the LogoUploader in 794542b1.

We have to stub calls to the Asset Manager API in the
ConsultationsControllerTest but otherwise no changes to the tests are
required.

We've verified in development that:
 - adding an attachement to a Consultation
 - removing an attachment from a Consultation
 - replacing an attachment on a Consultation

all result in the correct calls being made to Asset Manager. The
existing behaviour when a draft consultation is discarded is to not
remove the associated asset from disk. It is not clear whether this is
intended behaviour, but switching the storage engine preserves this
behaviour (that is, no DELETE call is made to the Asset Manager for
the associated asset).