Catch all kinds of 500 errors from GDS API
Currently we only catch 502 errors, and all others get sent
through to Sentry. Looking at the history of why we picked only
502 it seems to be because this is the code we get when Rummager
times out.

Since then we've integrated with Content Store which will often
timeout and come through to us as a 504 error. It doesn't seem
useful to log this as an exception since there is nothing we can
do about it in Whitehall.

It seems to me that not reporting any kind of external server error
seems like the right thing to do since it's not in our control and
increases the noise when looking through the exception log. Instead
this commit will mean that we only see client errors which is more
likely to expose a problem on our side.