Construct the URL of Whitehall assets served by Asset Manager
This is in preparation for using the `AssetManagerStorage` engine
instead of the `AssetManagerAndQuarantinedFileStorage` engine for
organisation logos.

Prior to this change, we were querying Asset Manager for the location of
the asset in order to return the URL. The asynchronous nature of
uploading the assets to Asset Manager meant that it was possible to ask
Asset Manager for the location of an asset that hadn't yet been
uploaded. This caused problems when trying to edit organisations with
custom logos when those logos hadn't yet been saved to Asset Manager.

We're avoiding the problem by manually constructing the URL using the
legacy_url_path and the known asset host. This essentially duplicates
the logic in Asset Manager's `file_url` key returned from
`AssetPresenter#as_json`[1]. Note that we're unable to use
`Plek#asset_root` because that reads from `GOVUK_ASSET_ROOT` and that's
set to 'whitehall-admin' for whitehall-admin[2]. Fortunately, we can use
`Plek#public_asset_host`[3] instead. It's possible we should update
`AssetPresenter` to use this method too but we're going to leave that
for another time.

[1]: https://github.com/alphagov/asset-manager/blob/bcc1b7085e9b050033a9d1227f81d873a7943822/app/presenters/asset_presenter.rb#L15
[2]:
https://github.com/alphagov/govuk-puppet/commit/960020ac00ced72a66806601ba2124266dad545b
[3]:
https://github.com/alphagov/plek/blob/4d63a574ab36e9e8833ff14471482d1cbb110e61/lib/plek.rb#L98