Add integration tests for ConsultationResponseFormData
We intend to use the `AssetManagerStorage` store for organisation logos
instead of the `AssetManagerAndQuarantinedFileStorage` store. When we
make this switch we will lose some coverage of the behaviour of the
`AssetManagerAndQuarantinedFileStorage` in these integration tests. We
don't currently have integration tests for the use of the
`AssetManagerAndQuarantinedFileStorage` for
`ConsultationResponseFormData` so adding some gives us confidence that
we can make the change we want.

In the `setup` of `RemovingAConsultationResponseFormData` test we are
setting an instance variable `@file_path` that is only used in the
first `test` block. However we noticed that if we do not make an
explicit call to `@consultation_response_form_data.file` before
calling `remove_file!` then `Services.asset_manager.delete_asset` is
not called. We're confident that this doesn't happen in reality but
has something to do with how CarrierWave loads uploaders on-demand. We
spent some time trying to fully understand what was causing this in
CarrierWave but gave up and have accepted that we need to make this
call to `file` in the `setup`.