Redirect deleted teacher supply stats announcement
The statistics announcement originally published at
/government/statistics/announcements/teacher-supply-model-2018-to-2019
was cancelled because the data was combined with the published statistics at
/government/statistics/tsm-and-initial-teacher-training-allocations-2018-to-2019.

The publishers want
/government/statistics/announcements/teacher-supply-model-2018-to-2019
to redirect to
/government/statistics/announcements/tsm-and-initial-teacher-training-allocations-2018-to-2019.

The content team attempted to unpublish and redirect the cancelled
announcement but for some reason that failed. It looked, upon initial
investigation, like the announcement was hard deleted from whitehall's
database. On further investigation, however, this was not actually the
case, it was just that the `default_scope` on the model was hiding the
data.

This migration creates a redirect for the content id of the now deleted
statistics announcement.