Treat ‘British Trade and Cultural Office’ as an embassy office type
The FCO have requested the ‘British nationals
should contact …’ text be removed from the
embassies index page for Taiwan:

https://www.gov.uk/world/embassies

Currently, this displays because Taiwan is one of
the SPECIAL_CASES in EmbassyPresenter. If we
remove it, the text is removed but so is the link.
We want to keep the link.

To do this, we need to add the office type of
Taiwan’s embassy office to the recognised types in
Embassy so it continues to be shown in the list.

Currently, this is the only office with this type
so it shouldn’t affect anything else.