Ensure old logos are removed when they're being replaced
We've seen an exception in production[1] when someone tried to replace
an organisation logo with a file of the same name. The file wasn't
removed from Asset Manager and so the replacement couldn't be uploaded.

I've replicated the problem in the AssetManagerIntegrationTest and fixed
it by adding the `AssetManagerStorage::File#path` method. I'm assuming
that CarrierWave checks whether the path has changed when deciding
whether it needs to upload a new file during replacement.

[1]: https://sentry.io/govuk/app-whitehall/issues/391658602/events/9208224189/