Only serve world locations that are delegations
This commit returns a not found error for any frontend requests for world locations unless they are international delegations, that are still served by `whitehall`. All other world location pages are served by `collections`. This stops serving pages for world locations that not enabled (such as the UK) and also requests for world locations with mixed-case base paths, which essentially bypasses `collections`.