Add reviewable
Shows a list of content items needing review (review by date expiring
in the next 7 days). Then links through to allow you update that
review date. All review date data is stored in the Content
Performance Manager.

And link to the edit functions for each item. This is part of a spike to
prototype "life expectancy" of content.