Upgrade many gems
It was noticed that we were using an outdated version of Rubocop
which had a security vulnerability. I decided to upgrade our
version of govuk-lint to get an updated version of Rubocop, but
that had a conflict with some other gems, so I had to upgrade the
rest of the gems at the same time.