Replace factory_girl gem with factory_bot
The FactoryGirl gem has been deprecated leading to the following
warning when running tests

> DEPRECATION WARNING: The factory_girl gem is deprecated. Please
> upgrade to factory_bot.

This commit replaces the `factory_girl` gem with the `factory_bot`
gem.

The change in version number in Gemfile.lock is slightly
confusing (from `4.9.0` to `4.8.2`) - this isn't a downgrade. `4.9.0`
of `factory_girl` was released with this deprecation warning as
the only change[1]. `4.8.2` of `factory_bot` is the latest version of
the new gem.

I've renamed the constants `FactoryGirl` to `FactoryBot` as per the
update instructions[2].

[1] https://github.com/thoughtbot/factory_bot/pull/1061
[2] https://github.com/thoughtbot/factory_bot/blob/4-9-0-stable/UPGRADE_FROM_FACTORY_GIRL.md