Do not return visible draft taxons from EditionTaxonsFetcher
The EditionTaxonsFetcher class is being used to populate the tagged taxons for an edition for the edition show page in the Whitehall Admin UI. We do not want to show invisible taxons as it would be confusing for publisher users. We filter out these invisible draft links by figuring out which of the tagged taxons are published and visible drafts and only show these.