Don't include singletons when disabling publishing api
XPath::HTML is now included in ObjectSpace, when doing
XPath::HTML == PublishesToPublishingApi it returns an XPath::Expression,
which is truthy. Therefore this is included inside the all_classes, and
as it is a singleton it causes a fatal error when trying to stub any
instance of it.

To fix this, we have ensured that no singletons will get included by the
select