Stop notifying Rummager for Finder Format
We no longer need to publish Finders directly to Rummager as this is now done through the notification queue via the publishing API.

The static pages which publish finders include indexable content which the publishing API does not pass through to Rummager. Only 3 of those Finder static pages had different indexable content to the description (which is passed through to Rummager) so we have decided to use the description field for indexable content as that contains nearly all the words that people use to get to those pages.