Fix document filter presenter previous/next URL
This commit adds a new previous/next URL to the document filter presenter that always points to the “web” (human-readable) version of the page rather than being context-driven. Currently, this is a JSON URL when called from the frontend JavaScript which means the previous/next buttons on the relevant finders point to JSON URLs.