Integration test for creating an ImageUploader image
The `People` model mounts the `ImageUploader` on `image` so we can use
that to integration test the use of the
`asset_manager_and_quarantined_file_storage` storage engine on
ImageUploader.

This test ensures that the correct number of asset versions (1 for the
original plus one for each `version` defined in `ImageUploader`) are
sent to Asset Manager when a new person is created with a bitmap
image. We also verify that the same files are stored in the
quarantined file store.