Redirect asset requests to asset host
It's currently possible to request Whitehall assets (and attachments) at
/government/uploads via both www.gov.uk and the asset host (e.g.
assets.publishing.gov.uk). We want all asset requests to go via the
asset host so that we can make changes to the config in a single place.

I'm using `Plek#public_asset_host` as it returns "The public-facing
asset base URL"[1]. I've tested this on integration and confirmed that
it returns
'https://assets-origin.integration.publishing.service.gov.uk'.

I initially tried to do the redirect in the routes file but it was
proving to be quite hard so I've gone with a method in the
`PublicUploadsController` instead.

I had to update upload_access_test to simulate making requests against
the asset host. Otherwise all the tests failed because they were being
redirect to the asset host.

Note that attachments are still available on both www.gov.uk and the
asset host. I expect we'll update this when we start working on the
migration of attachments to Asset Manager.

[1]:
https://github.com/alphagov/plek/blob/4d63a574ab36e9e8833ff14471482d1cbb110e61/lib/plek.rb#L97