Allow VirusScanHelpers.simulate_virus_scan to scan versions
Before this commit calling simulate_virus_scan only moved the original
version of an asset from the `incoming` directory to the `clean`
directory. This isn't quite the same as how virus scanning works in
production[1] in that it failed to move all of the generated
"versions" of an uploaded asset from `incoming` to `clean`. This
commit simulates the scanning of the original file and all its
versions if the `include_versions` option is truthy.

[1] Implemented using a cron job which calls https://github.com/alphagov/govuk-puppet/blob/eb86841c6fab1a9c30db4336aad2573ed2259055/modules/govuk/templates/node/s_asset_base/process-uploaded-attachments.sh.erb