Integration test for replacing an ImageUploader image
In ImageUploader we set the config parameter
`config.remove_previously_stored_files_after_update` to `false`. This
means that unlike the two other Uploaders that are tested in this
integration test, we do not expect the replaced images to be deleted
from the file system or asset manager. Testing that something doesn't
happen by expectation isn't ideal, but I have verified that setting the
config parameter mentioned above to `true` does cause the three test
cases to fail as expected.