Send empty array of people links to API if no roles attached
If a publisher removes people from an article, we stop sending the
“people” key in the links hash. Because this is processed on the
Publishing API side using PatchLinks, the people aren’t removed as it
ignores link types which aren’t included. If we want to remove all
people from a document, we need to send an empty array instead.

Fixes Zendesk ticket 2531517