MigrateAssetsToAssetManager#to_s returns a count
Before this commit MigrateAssetsToAssetManager#to_s returned a newline
separated list of all of the files to be migrated. When we migrate
`system/uploads/image_data/file`, for example, there will be almost
500,000 files[1]. It is probably not a good idea to generate this much
output in the rake task that calls this method, and a simple count
will be sufficient to verify that the correct number of assets have
been migrated.

I've also removed AssetFilePaths#join. As of this commit it is no
longer used.

[1] https://github.com/alphagov/asset-manager/issues/341