Rename extension to format in public_upload route segment
This `public_upload` route was unnecessarily complicated.

Rails will treat anything after the final period in a wildcard route as
the format[1] by default.  By adding the `:extension` dynamic segment we
were essentially duplicating that default behaviour. This meant that if
we were to request a file containing more than one period, e.g.
/government/uploads/path/to/file.name.png then `params[:path]` would
contain 'path/to/file', `params[:extension]` would contain 'name' and
`params[:format]` would contain 'png'.

This caused me a problem when trying to redirect asset requests to the
canonical asset host as calling `url_for(params)` (using the params
above) would end up dropping the format of the file.

I attempted to rename the `extension` segment in the other
asset/attachment routes but struggled to rename it for the attachment
preview route: In that case, the format is actually HTML even though the
extension will (always?) be CSV.

[1]:
http://guides.rubyonrails.org/routing.html#route-globbing-and-wildcard-segments