Use the asset host in consultation response form URLs
We want Whitehall assets to be served by the asset host (instead of
either by the asset host or gov.uk) so that we can make caching and
routing changes in a single place.

PR #3627 adds a redirect to ensure that asset requests made to gov.uk
are redirect to the asset host. This change will ensure that we're using
the canonical asset host when writing consultation content to the
Content Store.

These URLs are rendered in the "Ways to respond" section on a
consultation page.