Avoid redirecting requests for HMRC assets
We host a number of assets for HMRC's Basic PAYE Tools software[1] on
gov.uk. Some of these files are requested from within the Basic PAYE
Tools app - to check for updates, for example. It's not clear how the
app will behave if the request responds with a 302 redirect instead of a
200 with the content of the file. We plan to ask HMRC to update the URL
that they're checking from within the app but we should avoid
redirecting them in the short term to avoid any problems.

I initially used `=~` to check the regexp in the
`PublicUploadsController` but have changed it to `match?` as instructed
by RuboCop.

I explored reducing the duplication of 'government/uploads' in the app
but it became clear that it was going to take quite some time to do it.
It's not obvious that it's worth the investment at the moment given that
some of this code will go away as we migrate more assets to Asset
Manager.

[1]: https://www.gov.uk/basic-paye-tools